marketing-code-test/octopus/server/requirements.txt
aniso8601==3.0.2
asgiref==3.2.7
attrs==19.3.0
awscli==1.17.9
botocore==1.14.9
Click==7.0
colorama==0.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.2.0
docutils==0.15.2
factory-boy==2.12.0
Faker==4.0.2
graphene==2.1.7
graphene-django==2.9.0
graphql-core==2.3.1
graphql-relay==2.0.0
importlib-metadata==1.5.0
iso8601==0.1.12
jmespath==0.9.4
more-itertools==8.2.0
packaging==20.1
pbr==5.4.3
pip-tools==4.2.0
pluggy==0.13.1
promise==2.3
py==1.8.1
pyasn1==0.4.8
pyparsing==2.4.6
pytest==5.3.5
pytest-clarity==0.3.0a0
pytest-django==3.8.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.2
rsa==3.4.2
Rx==1.6.1
s3transfer==0.3.2
singledispatch==3.4.0.3
six==1.13.0
sqlparse==0.3.1
stevedore==1.31.0
termcolor==1.1.0
text-unidecode==1.3
urllib3==1.25.8
virtualenv==16.7.7
virtualenv-clone==0.5.3
virtualenvwrapper==4.8.4
wcwidth==0.1.8
zipp==2.1.0
